+++ app/graphql/types/admin_toolkit/cost_threshold_type.rb
@@ -4,7 +4,6 @@ module Types
   module AdminToolkit
     class CostThresholdType < BaseObject
       field :id, ID, null: true
-      field :not_exceeding, Float, null: true
       field :exceeding, Float, null: true
     end
   end