+++ db/migrate/20220213091922_add_new_column_to_projects.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class AddNewColumnToProjects < ActiveRecord::Migration[6.1]
+  def change
+    safety_assured do
+      add_column :projects, :exceeding_cost, :decimal
+    end
+  end
+end