+++ db/migrate/20220306212410_update_projects_pct_costs.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class UpdateProjectsPctCosts < ActiveRecord::Migration[6.1]
+  def up
+    change_column :projects_pct_costs, :payback_period, :float
+  end
+
+  def down
+    change_column :projects_pct_costs, :payback_period, :integer
+  end
+end