+++ db/structure.sql
@@ -502,7 +502,6 @@ CREATE TABLE public.admin_toolkit_competitions (
 
 CREATE TABLE public.admin_toolkit_cost_thresholds (
     id uuid DEFAULT public.gen_random_uuid() NOT NULL,
-    not_exceeding numeric,
     exceeding numeric,
     created_at timestamp(6) without time zone NOT NULL,
     updated_at timestamp(6) without time zone NOT NULL
@@ -858,7 +857,8 @@ CREATE TABLE public.projects (
     building_type character varying,
     cable_installations text[] DEFAULT '{}'::text[],
     priority_tac character varying,
-    access_technology_tac character varying
+    access_technology_tac character varying,
+    exceeding_cost numeric
 );
 
 
@@ -1041,13 +1041,13 @@ CREATE TABLE public.projects_pct_costs (
     arpu numeric(15,2),
     lease_cost numeric(15,2),
     penetration_rate double precision,
-    payback_period integer DEFAULT 0 NOT NULL,
     system_generated_payback_period boolean DEFAULT true NOT NULL,
     created_at timestamp(6) without time zone NOT NULL,
     updated_at timestamp(6) without time zone NOT NULL,
     build_cost numeric(15,2),
     roi numeric(15,2),
-    connection_cost_id uuid
+    connection_cost_id uuid,
+    payback_period double precision
 );
 
 
@@ -2305,6 +2305,8 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20211220101735'),
 ('20211220120052'),
 ('20211231081347'),
-('20220103082627');
+('20220103082627'),
+('20220213091922'),
+('20220306212410');