+++ spec/factories/admin_toolkit/competitions.rb
@@ -7,7 +7,6 @@ FactoryBot.define do
     factor { 0.9 }
     lease_rate { 75 }
     description { 'With only HFC of the EVU (Open Access Layer 1 & 2; tendency rather Salt & Sunrise)' }
-    calculation_type { 'Swisscom DSL' }
 
     trait :swisscom_dsl do
       name { 'Swisscom DSL' }