+++ spec/graphql/resolvers/admin_toolkit/cost_threshold_resolver_spec.rb
@@ -29,7 +29,7 @@ RSpec.describe Resolvers::AdminToolkit::CostThresholdResolver do
 
   def query
     <<~GQL
-      query { adminToolkitCostThreshold { exceeding id notExceeding } }
+      query { adminToolkitCostThreshold { exceeding id } }
     GQL
   end
 end